Create training composable components file segregation
