+++ app/src/main/java/com/example/stretchy/features/createtraining/ui/composable/buttons/AddOrSubstractButtons.kt
@@ -1,0 +1,56 @@
+package com.example.stretchy.features.createtraining.ui.composable
+
+import androidx.compose.foundation.layout.*
+import androidx.compose.material.Button
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun AddOrSubtractButtons(onTextEntered: (value: Int) -> Unit) {
+    val modifier = Modifier
+        .width(44.dp)
+        .padding(end = 6.dp)
+    Row {
+        Button(modifier = modifier,
+            contentPadding = PaddingValues(0.dp),
+            onClick = { onTextEntered(-10) })
+        {
+            Text(text = "-10")
+        }
+        Button(modifier = modifier,
+            contentPadding = PaddingValues(0.dp),
+            onClick = { onTextEntered(-5) })
+        {
+            Text(text = "-5")
+        }
+        Button(modifier = modifier,
+            contentPadding = PaddingValues(0.dp),
+            onClick = { onTextEntered(-1) })
+        {
+            Text(text = "-1")
+        }
+        Spacer(modifier = Modifier.width(52.dp))
+        Button(modifier = modifier,
+            contentPadding = PaddingValues(0.dp),
+            onClick = { onTextEntered(+1) })
+        {
+            Text(text = "+1")
+        }
+        Button(
+            modifier = modifier,
+            contentPadding = PaddingValues(0.dp),
+            onClick = { onTextEntered(+5) })
+        {
+            Text(text = "+5")
+        }
+        Button(
+            modifier = modifier,
+            contentPadding = PaddingValues(0.dp),
+            onClick = { onTextEntered(+10) })
+        {
+            Text(text = "+10")
+        }
+    }
+}
